+++ CHANGELOG.md
@@ -27,6 +27,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
   config setting.
 - Add way to create a relationship struct: `modelName.R.NewStruct()`
 - MySQL numeric, fixed, dec types now generate using the new types.(Null)Decimal
+- Use bindata as the default method of accessing templates, this prevents many
+  bug reports we've had in the past. During development or to otherwise opt out
+  the --basedir flag will not load from bindata (driver's are the exception).
 
 ### Changed